Parameterize boot dependencies
This allows the usage of variables defined in the `build.boot` file or
`profile.boot` to be used in defining dependencies. For example, the
version of Datomic can be set as a variable and used for both peer and
transactor dependencies.

This also adds an extra exclusion to the Datomic peer to prevent
annoying slf4j warnings.